app/requirements.txt
# app/requirements.txt

annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
async-typer==0.1.8
asyncpg==0.30.0
bcrypt==3.2.2
certifi==2025.1.31
cffi==1.17.1
click==8.1.8
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.11
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.1.2
markdown-it-py==3.0.0
mdurl==0.1.2
packaging==24.2
passlib==1.7.4
pillow==11.1.0
pyasn1==0.4.8
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
Pygments==2.19.1
PyPDF2==3.0.1
pytesseract==0.3.13
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
pyzbar==0.1.9
rich==13.9.4
rsa==4.9
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.39
starlette==0.46.1
typer==0.9.4
typing_extensions==4.12.2
uvicorn==0.34.0
